Separated out functions and tests
